IDE-3989 reconstruct service builder tests project
