+++ ui-tests/tools/com.liferay.ide.ui.service.builder/com.liferay.ide.ui.service.builder.editor/com.liferay.ide.ui.service.builder.editor.operation/build.properties
