+++ ui-tests/tools/com.liferay.ide.ui.service.builder/com.liferay.ide.ui.service.builder.editor/com.liferay.ide.ui.service.builder.editor.operation/pom.xml
@@ -20,13 +20,13 @@
 	<modelVersion>4.0.0</modelVersion>
 
 	<parent>
-		<groupId>com.liferay.ide.ui.tests</groupId>
-		<artifactId>tools-tests</artifactId>
+		<groupId>com.liferay.ide.ui.tests.service.builder.editor</groupId>
+		<artifactId>service-builder-editor-tests</artifactId>
 		<version>3.2.0-SNAPSHOT</version>
 	</parent>
 
-	<artifactId>com.liferay.ide.ui.service.builder.tests</artifactId>
+	<artifactId>com.liferay.ide.ui.service.builder.editor.operation.tests</artifactId>
 	<packaging>eclipse-test-plugin</packaging>
 
-	<name>Liferay IDE UI Service Builder Tests</name>
+	<name>Liferay IDE UI Service Builder Editor Operation Tests</name>
 </project>
