+++ ui-tests/tools/com.liferay.ide.ui.service.builder/com.liferay.ide.ui.service.builder.editor/com.liferay.ide.ui.service.builder.editor.operation/src/com/liferay/ide/ui/service/builder/tests/EditorServiceBuilderTests.java
@@ -16,59 +16,35 @@ package com.liferay.ide.ui.service.builder.tests;
 
 import com.liferay.ide.ui.liferay.SwtbotBase;
 import com.liferay.ide.ui.liferay.support.project.SdkProjectSupport;
-import com.liferay.ide.ui.liferay.support.server.PureTomcat62Support;
+import com.liferay.ide.ui.liferay.support.server.PureTomcat70Support;
 import com.liferay.ide.ui.liferay.util.RuleUtil;
 
 import org.junit.ClassRule;
-import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.RuleChain;
 
 /**
- * @author Joye Luo
  * @author Terry Jia
+ * @author Ying Xu
  */
-@Ignore("there are a few sub process still running after Refresh server adapter, need more research")
-public class NewServiceBuilderPortletSdk62Tests extends SwtbotBase {
+public class EditorServiceBuilderTests extends SwtbotBase {
 
-	public static PureTomcat62Support tomcat62 = new PureTomcat62Support(bot);
+	public static PureTomcat70Support tomcat = new PureTomcat70Support(bot);
 
 	@ClassRule
-	public static RuleChain chain = RuleUtil.getTomcat62SdkRuleChain(bot, tomcat62);
+	public static RuleChain chain = RuleUtil.getTomcat7xSdkRuleChain(bot, tomcat);
 
-	@Ignore("ignore as service builder in sdk62 is only able run in java 7")
 	@Test
-	public void buildServiceOnProject() {
-		if (envAction.notInternal()) {
-			return;
-		}
-
-		viewAction.switchLiferayPerspective();
-
-		wizardAction.openNewLiferayPluginProjectWizard();
-
-		wizardAction.newPlugin.prepareServiceBuilderPortletSdk(project.getNamePortlet());
-
-		wizardAction.finish();
-
-		jobAction.waitForIvy();
-
-		jobAction.waitForValidate(project.getNamePortlet());
-
-		viewAction.project.runBuildServices(project.getNamePortlet());
-
-		jobAction.waitForConsoleContent("build.xml", "BUILD SUCCESSFUL", 30 * 1000);
-
-		viewAction.project.closeAndDelete(project.getNamePortlet());
+	public void addColumn() {
 	}
 
 	@Test
-	public void buildWSDDOnProject() {
-		if (envAction.notInternal()) {
-			return;
-		}
+	public void addEntity() {
+	}
 
+	@Test
+	public void createServiceBuilderPortlet() {
 		viewAction.switchLiferayPerspective();
 
 		wizardAction.openNewLiferayPluginProjectWizard();
@@ -81,9 +57,13 @@ public class NewServiceBuilderPortletSdk62Tests extends SwtbotBase {
 
 		jobAction.waitForValidate(project.getNamePortlet());
 
-		viewAction.project.runBuildWSDD(project.getNamePortlet());
+		viewAction.project.openFile(project.getNamePortlet(), "docroot", "WEB-INF", "service.xml");
 
-		jobAction.waitForConsoleContent("build.xml", "BUILD SUCCESSFUL", 300 * 1000);
+		editorAction.serviceXml.switchTabDiagram();
+
+		editorAction.serviceXml.switchTabOverview();
+
+		editorAction.serviceXml.switchTabSource();
 
 		viewAction.project.closeAndDelete(project.getNamePortlet());
 	}
