+++ ui-tests/tools/com.liferay.ide.ui.service.builder/com.liferay.ide.ui.service.builder.editor/pom.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+/*******************************************************************************
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ *
+ *******************************************************************************/
+-->
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<parent>
+		<groupId>com.liferay.ide.ui.tests.service.builder</groupId>
+		<artifactId>service-builder-tests</artifactId>
+		<version>3.2.0-SNAPSHOT</version>
+	</parent>
+
+	<groupId>com.liferay.ide.ui.tests.service.builder.editor</groupId>
+	<artifactId>service-builder-editor-tests</artifactId>
+	<packaging>pom</packaging>
+
+	<name>Liferay IDE UI Service Builder Editor</name>
+
+	<modules>
+		<module>com.liferay.ide.ui.service.builder.editor.operation</module>
+	</modules>
+</project>
