+++ ui-tests/tools/com.liferay.ide.ui.service.builder/com.liferay.ide.ui.service.builder.wizard/com.liferay.ide.ui.service.builder.wizard.create/.gitignore
@@ -1,0 +1,5 @@
+/screenshots
+/.settings
+/.classpath
+/target
+/.project
