+++ ui-tests/tools/com.liferay.ide.ui.service.builder/com.liferay.ide.ui.service.builder.wizard/com.liferay.ide.ui.service.builder.wizard.create/src/com/liferay/ide/ui/service/builder/tests/NewServiceBuilderPortletSdk62Tests.java
@@ -16,34 +16,35 @@ package com.liferay.ide.ui.service.builder.tests;
 
 import com.liferay.ide.ui.liferay.SwtbotBase;
 import com.liferay.ide.ui.liferay.support.project.SdkProjectSupport;
-import com.liferay.ide.ui.liferay.support.server.PureTomcat70Support;
+import com.liferay.ide.ui.liferay.support.server.PureTomcat62Support;
 import com.liferay.ide.ui.liferay.util.RuleUtil;
 
 import org.junit.ClassRule;
+import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.RuleChain;
 
 /**
+ * @author Joye Luo
  * @author Terry Jia
+ * @author Ying Xu
  */
-public class EditorServiceBuilderTests extends SwtbotBase {
+@Ignore("there are a few sub process still running after Refresh server adapter, need more research")
+public class NewServiceBuilderPortletSdk62Tests extends SwtbotBase {
 
-	public static PureTomcat70Support tomcat = new PureTomcat70Support(bot);
+	public static PureTomcat62Support tomcat62 = new PureTomcat62Support(bot);
 
 	@ClassRule
-	public static RuleChain chain = RuleUtil.getTomcat7xSdkRuleChain(bot, tomcat);
+	public static RuleChain chain = RuleUtil.getTomcat62SdkRuleChain(bot, tomcat62);
 
+	@Ignore("ignore as service builder in sdk62 is only able run in java 7")
 	@Test
-	public void addColumn() {
-	}
+	public void buildServiceOnProject() {
+		if (envAction.notInternal()) {
+			return;
+		}
 
-	@Test
-	public void addEntity() {
-	}
-
-	@Test
-	public void createServiceBuilderPortlet() {
 		viewAction.switchLiferayPerspective();
 
 		wizardAction.openNewLiferayPluginProjectWizard();
@@ -56,13 +57,34 @@ public class EditorServiceBuilderTests extends SwtbotBase {
 
 		jobAction.waitForValidate(project.getNamePortlet());
 
-		viewAction.project.openFile(project.getNamePortlet(), "docroot", "WEB-INF", "service.xml");
+		viewAction.project.runBuildServices(project.getNamePortlet());
 
-		editorAction.serviceXml.switchTabDiagram();
+		jobAction.waitForConsoleContent("build.xml", "BUILD SUCCESSFUL", 30 * 1000);
 
-		editorAction.serviceXml.switchTabOverview();
+		viewAction.project.closeAndDelete(project.getNamePortlet());
+	}
 
-		editorAction.serviceXml.switchTabSource();
+	@Test
+	public void buildWSDDOnProject() {
+		if (envAction.notInternal()) {
+			return;
+		}
+
+		viewAction.switchLiferayPerspective();
+
+		wizardAction.openNewLiferayPluginProjectWizard();
+
+		wizardAction.newPlugin.prepareServiceBuilderPortletSdk(project.getNamePortlet());
+
+		wizardAction.finish();
+
+		jobAction.waitForIvy();
+
+		jobAction.waitForValidate(project.getNamePortlet());
+
+		viewAction.project.runBuildWSDD(project.getNamePortlet());
+
+		jobAction.waitForConsoleContent("build.xml", "BUILD SUCCESSFUL", 300 * 1000);
 
 		viewAction.project.closeAndDelete(project.getNamePortlet());
 	}
