+++ ui-tests/tools/com.liferay.ide.ui.service.builder/com.liferay.ide.ui.service.builder.wizard/com.liferay.ide.ui.service.builder.wizard.create/src/com/liferay/ide/ui/service/builder/tests/NewServiceBuilderPortletSdkTests.java
@@ -27,6 +27,7 @@ import org.junit.rules.RuleChain;
 /**
  * @author Joye Luo
  * @author Terry Jia
+ * @author Ying Xu
  */
 public class NewServiceBuilderPortletSdkTests extends SwtbotBase {
