+++ ui-tests/tools/com.liferay.ide.ui.service.builder/pom.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+/*******************************************************************************
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ *
+ *******************************************************************************/
+-->
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<parent>
+		<groupId>com.liferay.ide.ui.tests</groupId>
+		<artifactId>tools-tests</artifactId>
+		<version>3.2.0-SNAPSHOT</version>
+	</parent>
+
+	<groupId>com.liferay.ide.ui.tests.service.builder</groupId>
+	<artifactId>service-builder-tests</artifactId>
+	<packaging>pom</packaging>
+
+	<name>Liferay IDE UI Service Builder</name>
+
+	<modules>
+		<module>com.liferay.ide.ui.service.builder.wizard</module>
+		<module>com.liferay.ide.ui.service.builder.editor</module>
+	</modules>
+</project>
