+++ ui-tests/tools/pom.xml
@@ -60,7 +60,7 @@
 		<module>com.liferay.ide.ui.sdk.tests</module>
 		<module>com.liferay.ide.ui.server</module>
 		<module>com.liferay.ide.ui.workspace.tests</module>
-		<module>com.liferay.ide.ui.service.builder.tests</module>
+		<module>com.liferay.ide.ui.service.builder</module>
 		<module>com.liferay.ide.ui.service.builder.deploy.tests</module>
 		<module>com.liferay.ide.ui.kaleo.tests</module>
 		<module>com.liferay.ide.ui.kaleo.deploy.tests</module>
